requirements.txt
fastapi>=0.104.0,<0.105.0
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.1.0,<2.2.0
uvicorn>=0.24.0,<0.25.0
SQLAlchemy>=2.0.0,<2.1.0
alembic>=1.12.0,<1.13.0
email-validator==2.1.0.post1
PyJWT==2.8.0
python-multipart==0.0.6
python-decouple>=3.8.0,<3.9.0 #read the environment variables
passlib>=1.7.4,<1.8.0
emails==0.6
asyncpg>=0.29.0,<0.30.0

# test dependencies
flake8==6.1.0
isort==5.12.0
pytest>=7.4.0,<7.5.0
pytest-mock>=3.12.0,<3.13.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-env>=1.1.0,<1.2.0
dotmap==1.3.30
